ui change

+++ src/components/back/look.vue
@@ -1,6 +1,8 @@
 <template>
   <div class="look">
       <hd></hd>
+      <h1 class="pull-left">随想管理   主页面</h1>
+      <h1 class="pull-right"><router-link to="/insert">新增</router-link></h1>
       <el-table :data="articles">
         <el-table-column prop="time" label="日期" width="140">
         </el-table-column>
@@ -39,7 +41,6 @@ export default {
       //一般fetch方法就是获取数据的
       this.$http.get('/articles').then(res => {
         //看是否登录过期
-        console.log(res);
         if (res.status === 403){
           this.$message({
               showClose: true,
@@ -53,29 +54,34 @@ export default {
     },
 
     handleEdit(index, row) {
-      console.log(index);
-      console.log(row);
       this.$router.push(`/edit/${row._id}`);
     },
     handleDelete(index, row){
-      console.log(index);
-      console.log(row);
-      
-      //使用ES6 模板字符串 ` 这个符号
-      this.$http.delete(`/articles/${row._id}`).then(res =>{
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        //使用ES6 模板字符串 ` 这个符号
+        this.$http.delete(`/articles/${row._id}`).then(res =>{
+          this.$message({
+                showClose: true,
+                message: '删除成功！'
+          });
+          this.fetch();
+        });  
+        //按照id去删除，这个id是建立数据的时候，mongo数据库给分配的id。
+
+      }).catch(() => {
         this.$message({
-              showClose: true,
-              message: '删除成功！'
-        });
-        this.fetch();
-      });  
-      //按照id去删除，这个id是建立数据的时候，mongo数据库给分配的id。
+          type: 'info',
+          message: '已取消删除'
+        });          
+      });
     
     }
   },
   created() {
-
-    //if (this.$store.getters.getToken){
     this.fetch();
   },
   components:{
